Add functionality for page load + refactoring.
Get quiz data and create elements on page load. Now, when we access homepage.html, a call is made to the API to get all quizzes, and then for each quiz, an element is created inside the quizzes container.

Refactoring mainly in the script.js. Removed button and listener and changed the api call method from async to Promises, as this simplifies the code and is more than enough for our use case.

Also created simple implementation for the quiz description page, to test parameter passing using query strings between 2 pages.