Fix quiz description bug + Script Refactor
The bug where the quiz description wasn't showing sometimes when navigating to the quiz description page is now fixed. During the course of debugging, many different approaches where tried which led to code refactoring, which in turn improves code readability and performance. The changes were:

- Change getQuizzes to an async function: This led to the simplification of the init function. making the code more readable.
Code was refactored to achieve the following
- Quizzes in the displayQuizzes function are given as a parameter, instead of needing to load quizzes from storage. This is a better approach since the quiz data is available when the displayQuizzes function is called.
- Changed the way the selected quiz id is stored. Before, it was stored in the anchor element's id attribute, which is a bad practice. It is now stored in a data attribute of the element.
- Modified the listeners for the quiz links: Before, all anchor elements were selected and a listener was attached to them. Now, only the quiz anchor elements are selected for adding a listener.

The last part is probably what fixed the quiz description not showing, as the bug wasn't reproducible after this change.