Refactor frontend and add username login.
Moved all quiz/question logic from the frontend to the backend. This removes sensitive information (like questions and answers) from the session storage.

Added a starting page which simulates a user login. In this case, the user selects their username, and depending on the role (teacher or student), they 're redirected to either the teacher dashboard or the quizzes list page.