Fix quiz flow.
Major refactoring to the frontend javascript script, in order to fix the quiz flow.

Moving the logic from frontend to backend required additional steps that, when implemented, weren't compatible with the existing functionality.

As a result, the whole quiz flow logic needed to be rewritten, and extensive debugging was required, especially inside the question page.